Move fetch_orchestrator.py to fetching package
